Fixed test and import
