+++ lib/meta_strategy/web_meta_strategy.dart
@@ -1,4 +1,3 @@
-
 import 'dart:html';
 import 'dart:ui';
