+++ lib/screens/screens.dart
@@ -3,3 +3,4 @@ export 'register.dart';
 export 'not_found.dart';
 export 'profile/profile.dart';
 export 'privacy_policy/privacy_policy.dart';
+export 'home/home_page.dart';
